fix(thermo): address D4 pre-PR ultra-review (7 confirmed findings)
A local 8-dimension ultra-review (fan-out → adversarial-verify) surfaced 7 real,
reproduced findings (12 weaker ones refuted). All fixed:

HIGH
- Kramers test was seed-fragile — the headline Arrhenius-slope gate passed only
  on a lucky SEED_BASE (~50% of seeds failed). Now average each temperature's
  rate over N_REP seeds, restrict to the Arrhenius-valid regime (ΔV/kT≳1.3), and
  make the PRIMARY gate the per-kT match to analytic kramers_rate (±30%); the
  slope is a coarse ±25% sanity. Honest numbers in the recon (slope ≈ −3.4, the
  residual is the formula's moderate-barrier correction).
- Cold single-well calibration was corrupted: two-means split a lone Gaussian
  into two fake wells, biasing ⟨δx²⟩ ~3× low and fabricating switches — exactly
  the kT_eff calibration regime. Add a unimodality gate (bimodal flag + single-
  well path measuring ⟨δx²⟩ about the global mean, unbiased) + a cold-signal test.

MEDIUM
- Wrong physics in docs/comments: the hot ⟨δx²⟩ suppression is a measurement-gate
  truncation artifact, not anharmonicity (which actually *enhances* it, per the
  cold ratio ≈1.13). Corrected in kramers_pbit, the kt_eff docstring, and recon.
- analyze_channel panicked on ragged CSV rows — parse_csv now enforces rectangular
  data (clean ParseError), and the guard checks every sample, not just the first.
- t_us was u32 (wraps at 71.6 min) — too short for cold multi-hour captures.
  Widened LogRecord.t_us to u64 before the wire contract is hardware-locked.
- Recon S3 overclaimed a γ-fit/overlay/gap that ship at G5, and the CLI dangled at
  an uncallable kt_eff(). Demoted the recon scope to honest G1-vs-G5, surfaced the
  bimodal mode in the CLI, and fixed the prose.

New tests: cold_single_well_recovers_unbiased_variance, parse_csv_rejects_ragged_rows,
bimodal assertions. All crates clippy + fmt clean; heavy kramers test green.

+++ firmware/pbit-fw-core/src/record.rs
@@ -18,8 +18,10 @@ use crate::MAX_HALL;
 /// counts (one per p-bit); `ax/ay/az` are the base accelerometer ADC counts (the
 /// measured bath). The matching header comes from [`csv_header`](Self::csv_header).
 pub struct LogRecord {
-    /// Microseconds since acquisition start.
-    pub t_us: u32,
+    /// Microseconds since acquisition start. `u64` so it never wraps — a `u32`
+    /// microsecond counter rolls over at ~71.6 min and the cold-calibration /
+    /// rare-escape regime needs multi-hour captures.
+    pub t_us: u64,
     /// Commanded drive, scaled to `i16` (`-1..=1` ↔ `-32767..=32767`).
     pub drive: i16,
     /// Hall-sensor readings (raw ADC counts), one per p-bit.
